pyproject.toml
[project]
name = "ai-runny"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "garminconnect>=0.2.38",
    "openai>=1.40",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "matplotlib>=3.9",
    "streamlit>=1.45",
]
